+++ support/apps/apps.go
@@ -1,6 +1,3 @@
-//go:build !windows
-// +build !windows
-
 // Package apps provides methods for obtaining version information about various tools
 package apps