+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `[env]` Fixed compatibility with Windows
 - `[support]` Add basic support of collecting info on Windows
+- `[support/apps]` Added Windows support
 
 ### [13.0.0](https://kaos.sh/ek/13.0.0)